# Improve performance of fallback implementation