--- a/src/decode.rs
+++ b/src/decode.rs
@@ -95,12 +95,9 @@ pub fn hex_check(src: &[u8]) -> bool {
 }
 
 pub fn hex_check_fallback(src: &[u8]) -> bool {
-    for byte in src {
-        match byte {
-            b'A'..=b'F' | b'a'..=b'f' | b'0'..=b'9' => continue,
-            _ => {
-                return false;
-            }
+    for &byte in src {
+        if UNHEX[byte as usize] == NIL {
+            return false;
         }
     }
     true