--- a/src/decode.rs
+++ b/src/decode.rs
@@ -208,7 +208,7 @@ unsafe fn hex_decode_avx2(mut src: &[u8], mut dst: &mut [u8]) {
 }
 
 pub fn hex_decode_fallback(src: &[u8], dst: &mut [u8]) {
-    for (slot, bytes) in dst.iter_mut().zip(src.chunks(2)) {
+    for (slot, bytes) in dst.iter_mut().zip(src.chunks_exact(2)) {
         let a = unhex_a(bytes[0] as usize);
         let b = unhex_b(bytes[1] as usize);
         *slot = a | b;