Before:

```
bench_check_fallback    time:   [174.52 ns 175.79 ns 177.11 ns]

bench_faster_hex_encode_fallback                                                                            
                        time:   [146.59 ns 148.12 ns 149.78 ns]

bench_faster_hex_decode_fallback                                                                            
                        time:   [123.82 ns 125.09 ns 126.55 ns]
```

After:

```
bench_check_fallback    time:   [80.902 ns 81.617 ns 82.401 ns]                                 
                        change: [-54.745% -54.078% -53.337%] (p = 0.00 < 0.05)
                        Performance has improved.

bench_faster_hex_encode_fallback                                                                            
                        time:   [111.62 ns 112.84 ns 113.99 ns]
                        change: [-26.126% -24.597% -23.048%] (p = 0.00 < 0.05)
                        Performance has improved.

bench_faster_hex_decode_fallback                                                                            
                        time:   [108.05 ns 109.04 ns 110.02 ns]
                        change: [-13.733% -12.546% -11.147%] (p = 0.00 < 0.05)
                        Performance has improved.
```